// src/app/shared/models/classes/itsrequestfor.ts
export class Itsrequestfor {
    requestFor: string;
    Category?: string;
    sapModule?: string;
    SocialMedia?: string;
    VideoStreaming?: string;
    GuestName?: string;
    mobileBrandNdModel?: string;
    expectDeliveryDate?: string;
    expectReturnDate?: string;
    mobileSIMNO?: string;
    issueDate?: string;
    actualDeliveryDate?: string;
    ExpiryDate?: string;

}